Completed the full screen fluid box model
Was unable to use the flexible box model of CSS 3, it has changed so
much it's hard to find good documentation on it's actual implementation
in various browsers. Did instead handle resizing of the content using
JavaScript. Added a status bar to the bottom for status text while doing
operations.