Added support for searching for items to add.
Added the first Amazon Web Service integration which retreives movies
based on the user search input. The id and secret are stored in a
separate InTheBoks.ini for security reasons. Make sure you change the
code and config before production deployment, do not put the .ini file
on the server as this is downloadable. Add the keys to your web.config.