More features, more UI updates
Item selection is now complete, with overlay icon and border. Title are
displayed on hover.
Added support for CTRL-A and CTRL-C, to do selection of all the items in
a catalog and to "virtually" copy the selected items. This makes
InTheBoks act more like a real application.
Action buttons moved to a permanent toolbar that does not scroll with
the content.